Moving NavigationMapViewDelegate type into its own file. Also obsoleting delegate methods that are no longer being forwarded from the MGLMapView via its delegate.
